+++ code/qcommon/q_shared.c
@@ -58,7 +58,7 @@ const char *GetStringForID( stringID_table_t *table, int id ) {
 	return NULL;
 }
 
-static unsigned long holdrand = 0x89abcdef;
+static uint32_t holdrand = 0x89abcdef;
 
 int irand(int min, int max)
 {
@@ -67,7 +67,7 @@ int irand(int min, int max)
 	assert((max - min) < 32768);
 
 	max++;
-	holdrand = (holdrand * 214013L) + 2531011L;
+	holdrand = (holdrand * 214013) + 2531011;
 	result = holdrand >> 17;
 	result = ((result * (max - min)) >> 15) + min;
 	return(result);
